compile fix for standard sdk
ae66fad804002ea7588d7b20e3e09bf7842b0f9b introduced including winsock2.h
Unfortunately some flags are different when getservbyname gets declared
there on Windows Mobile compared to Windows CE standard SDK.

As there is no consistent way to resolve this and the only dependency
is on the servent struct, declare it in the header as well.

+++ src/time.c
@@ -492,7 +492,7 @@ void UnixTimeToFileTime(time_t t, LPFILETIME pft)
 
 	ll = Int32x32To64(t, 10000000) + 116444736000000000;
 	pft->dwLowDateTime = (DWORD)ll;
-	pft->dwHighDateTime = ll >> 32;
+    pft->dwHighDateTime = (DWORD)(ll >> 32);
 }
 
 // Returns to day of the year (0..365)
